Early transform of Config class in tests
DDSpecification requires transforming the Config class to make its
INSTANCE field non-final and volatile. This will fail in the presence of
some testing tooling (e.g. pitest). By moving this instrumentation to a
Spock global extension, it runs at an earlier stage, fixing the problem.
