+++ utils/test-utils/src/main/resources/META-INF/services/org.spockframework.runtime.extension.IGlobalExtension
@@ -1,0 +1,1 @@
+datadog.trace.test.util.ConfigTransformSpockExtension
